+++ packages/codesandbox/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.5.37
+
+### Patch Changes
+
+- Updated dependencies [3c4e595]
+  - computesdk@2.4.0
+  - @computesdk/provider@1.0.29
+
 ## 1.5.36
 
 ### Patch Changes