+++ packages/daytona/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @computesdk/daytona
 
+## 1.7.17
+
+### Patch Changes
+
+- Updated dependencies [3c4e595]
+  - computesdk@2.4.0
+  - @computesdk/provider@1.0.29
+
 ## 1.7.16
 
 ### Patch Changes