+++ packages/modal/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @computesdk/modal
 
+## 1.8.32
+
+### Patch Changes
+
+- Updated dependencies [3c4e595]
+  - computesdk@2.4.0
+  - @computesdk/provider@1.0.29
+
 ## 1.8.31
 
 ### Patch Changes