+++ packages/provider/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @computesdk/provider
 
+## 1.0.29
+
+### Patch Changes
+
+- Updated dependencies [3c4e595]
+  - computesdk@2.4.0
+
 ## 1.0.28
 
 ### Patch Changes